// libs/core/models/anime/anime-source.ts
/** Anime source. */
export enum AnimeSource {
	FourKomaMang = 'Four Koma Manga',
	Book = 'Book',
	CardGame = 'Card Game',
	LightNovel = 'Light Novel',
	Manga = 'Manga',
	MixedMedia = 'Mixed Media',
	Music = 'Music',
	Novel = 'Novel',
	Original = 'Original',
	PictureBook = 'Picture Book',
	Radio = 'Radio',
	VisualNovel = 'Visual Novel',
	WebManga = 'Web Manga',
	WebNovel = 'Web Novel',
	Other = 'Other',
	Unknown = 'Unknown',
	Game = 'Game',
}